Add uncertaintySizePalette in idd. Add Petals, BullEye, BoxWhisker in idd. Add scroll when left panel is overflow.
