+++ test/manual/mainGlobal.ts
@@ -18,9 +18,10 @@ module main {
                 x: [1, 2, 3, 4],
                 y: [1, 4, 6, 9],
                 shape: Plot.MarkerShape.Cross,
-                size: 8,
-                color: "red",
-                displayName: "mark"
+                size: [1,2,3,4],
+                color: [1, 2, 3, 4],
+                colorPalette: 'green, blue',
+                displayName: "mark",
             }),
             "elem3": Plot.heatmap({
                 x: [-1, 0, 1, -1, 0, 1, -1, 0, 1],
